feat(deps): mux => echo
We're moving to echo for a few reasons:
- JWT / user auth will be easier
- mux is unmaintained and I really need to learn another router/stuff
- Code is cleaner with echo
- We can still use gorilla/websocket with ech

+++ models/image.go
@@ -4,5 +4,5 @@ type Image struct {
 	Id         int64     `json:"id" db:"id"`
 	Date       Timestamp `json:"date" db:"created_at"`
 	Unattended bool      `json:"unattended" db:"unattended"`
-	EventId    int64     `json:"-" db:"event_id"`
+	EventId    int       `json:"-" db:"event_id"`
 }
